Working hamburger menu
Only has basic styling currently
