+++ src/components/header/style.sss
@@ -19,7 +19,7 @@
     white-space: nowrap
     overflow: hidden
 
-.back, #navigationDrawerButton
+.back
   outline: none
   border-radius: 50%
   margin: 5px 7px
@@ -36,47 +36,3 @@
     display: block
     margin: 8px
     height: 100%
-
-.navigationDrawer
-  visibility: hidden
-  opacity: 0
-
-.navigationDrawerContent, .navigationDrawerBackground
-  position: fixed
-  top: 0
-  left: 0
-
-.navigationDrawerBackground
-  width: 100%
-  height: 100%
-  background-color: rgba(#000, 0.2)
-
-.navigationDrawerContent
-  width: 320px
-  height: 100%
-  background-color: white
-  box-shadow: #000 0px 0px 15px
-  color: black
-
-  & > a
-    text-decoration: none
-    color: inherit
-
-  & li
-    width: 100%
-    height: 48px
-    padding: 0 16px
-    list-style-type: none
-    align-items: center
-    display: flex
-
-    & > *
-      display: inline-block
-
-    & > .icon
-      width: 56px
-
-.open
-  visibility: visible
-  opacity: 1
-
