+++ src/components/header/style.sss.d.ts
@@ -1,8 +1,2 @@
 export const header: string;
-export const navigationDrawerButton: string;
 export const back: string;
-export const navigationDrawer: string;
-export const navigationDrawerBackground: string;
-export const navigationDrawerContent: string;
-export const icon: string;
-export const open: string;
