+++ src/components/icon/style.sss
@@ -3,3 +3,4 @@
   height: 24px
   align-self: center
   justify-self: center
+  display: block
