+++ src/routes/home/style.sss.d.ts
@@ -6,3 +6,11 @@ export const cmp: string;
 export const pre: string;
 export const off: string;
 export const adminPanel: string;
+export const navigationDrawerButtonColumn: string;
+export const navigationDrawerButton: string;
+export const navigationDrawer: string;
+export const navigationDrawerBackground: string;
+export const navigationDrawerContent: string;
+export const icon: string;
+export const subheader: string;
+export const open: string;
